Session review fixes: live send publishes, library revision, stop race (#16)
- Live `send` publishes the state before it returns, so `state()` right
  after `send` shows what the control side applied, as docs/app-api.md says.
  Before, it showed the old state until the control thread's next pass.
- `library_list()` carries the revision its entries are. The library and its
  revision are published together. `state.library` (revision, count,
  position, pending) now describes the published list. Before, while
  indexing, the list was labelled with a newer revision than its entries.
  Client changes (a file added, a style that fails to load) publish at once.
- `stop()` claims the stop flag before taking the threads. Otherwise a
  concurrent second `stop` could take them and return without joining.
- `CyclePadPage` takes any i8 delta. It overflowed in debug and wrapped to
  the wrong page in release.
- Tests: every Launchkey pad, button and fader button against its AppCmd
  (state and output, three starting states), and pad page cycling with
  extreme deltas. Library revision consistency while indexing, plus a live
  send/state check and a concurrent stop.

+++ src/launchkey.rs
@@ -116,8 +116,8 @@ impl Page {
 
     /// The page `d` steps away, wrapping around (the terminal's Tab / Shift+Tab).
     pub fn cycle(self, d: i8) -> Page {
-        let n = Page::ALL.len() as i8;
-        Page::from_u8((self as i8 + d).rem_euclid(n) as u8)
+        let n = Page::ALL.len() as i16;
+        Page::from_u8((self as i16 + d as i16).rem_euclid(n) as u8)
     }
 
     /// The page's LED identity: RGB for the pads, plus bright and dim palette colours.
